+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 - Do not register `resources` directory for OOTB extensions in the readonly mode [#1601](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1601)
 - Adjusted default project & application settings [#1602](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1602)
 
+### `FlexibleSearch` enhancements
+- Added support of the aliases in the `Order` clause [#1605](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1605)
+
 ### `CCv2` enhancements
 - Added `S-User` alias for service `modified by` characteristic [#1596](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1596)